Move things back into hook
