+++ sections/exchange/MobileSwap/MobileSwap.tsx
@@ -1,21 +1,14 @@
 import React from 'react';
 import styled from 'styled-components';
-import useExchange from '../hooks/useExchange';
 import RatioSelect from './RatioSelect';
 import SwapInfoBox from './SwapInfoBox';
 import SwapInput from './SwapInput';
 import { SwapCurrenciesButton } from 'styles/common';
 import ArrowIcon from 'assets/svg/app/arrow-down.svg';
+import { useExchangeContext } from 'contexts/ExchangeContext';
 
 const MobileSwap: React.FC = () => {
-	const { handleCurrencySwap } = useExchange({
-		showPriceCard: true,
-		showMarketDetailsCard: true,
-		footerCardAttached: false,
-		routingEnabled: true,
-		persistSelectedCurrencies: true,
-		showNoSynthsCard: false,
-	});
+	const { handleCurrencySwap } = useExchangeContext();
 
 	return (
 		<MobileSwapContainer>
