+++ sections/exchange/TradeCard/Cards/BaseCurrencyCard.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { useTranslation } from 'react-i18next';
-import { wei } from '@synthetixio/wei';
-import { DEFAULT_CRYPTO_DECIMALS } from 'constants/defaults';
 import { useExchangeContext } from 'contexts/ExchangeContext';
 import SelectCurrencyModal from 'sections/shared/modals/SelectCurrencyModal';
 
@@ -13,7 +11,6 @@ import {
 	quoteCurrencyKeyState,
 	currencyPairState,
 } from 'store/exchange';
-import { truncateNumbers } from 'utils/formatters/number';
 import CurrencyCard from '../CurrencyCard';
 import { CurrencyKey } from 'constants/currency';
 
@@ -25,13 +22,12 @@ const BaseCurrencyCard: React.FC<BaseCurrencyCardProps> = ({ allowBaseCurrencySe
 	const { t } = useTranslation();
 	const baseCurrencyKey = useRecoilValue(baseCurrencyKeyState);
 	const quoteCurrencyKey = useRecoilValue(quoteCurrencyKeyState);
-	const [baseCurrencyAmount, setBaseCurrencyAmount] = useRecoilState(baseCurrencyAmountState);
+	const baseCurrencyAmount = useRecoilValue(baseCurrencyAmountState);
 	const setQuoteCurrencyAmount = useSetRecoilState(quoteCurrencyAmountState);
 	const setCurrencyPair = useSetRecoilState(currencyPairState);
 
 	const {
 		txProvider,
-		inverseRate,
 		baseCurrencyBalance,
 		openModal,
 		setOpenModal,
@@ -40,8 +36,9 @@ const BaseCurrencyCard: React.FC<BaseCurrencyCardProps> = ({ allowBaseCurrencySe
 		routeToMarketPair,
 		basePriceRate,
 		allTokensMap,
-		exchangeFeeRate,
 		oneInchQuoteQuery,
+		onBaseCurrencyAmountChange,
+		onBaseBalanceClick,
 	} = useExchangeContext();
 
 	return (
@@ -52,35 +49,9 @@ const BaseCurrencyCard: React.FC<BaseCurrencyCardProps> = ({ allowBaseCurrencySe
 				currencyName={baseCurrencyKey ? allTokensMap[baseCurrencyKey]?.name : null}
 				disabled={txProvider !== 'synthetix'}
 				amount={baseCurrencyAmount}
-				onAmountChange={async (value) => {
-					if (value === '') {
-						setBaseCurrencyAmount('');
-						setQuoteCurrencyAmount('');
-					} else {
-						setBaseCurrencyAmount(value);
-						if (txProvider === 'synthetix' && baseCurrencyKey != null) {
-							const quoteCurrencyAmountNoFee = wei(value).mul(inverseRate);
-							const fee = quoteCurrencyAmountNoFee.mul(exchangeFeeRate ?? 0);
-							setQuoteCurrencyAmount(
-								truncateNumbers(quoteCurrencyAmountNoFee.add(fee), DEFAULT_CRYPTO_DECIMALS)
-							);
-						}
-					}
-				}}
+				onAmountChange={onBaseCurrencyAmountChange}
 				walletBalance={baseCurrencyBalance}
-				onBalanceClick={async () => {
-					if (baseCurrencyBalance != null) {
-						setBaseCurrencyAmount(truncateNumbers(baseCurrencyBalance, DEFAULT_CRYPTO_DECIMALS));
-
-						if (txProvider === 'synthetix') {
-							const baseCurrencyAmountNoFee = baseCurrencyBalance.mul(inverseRate);
-							const fee = baseCurrencyAmountNoFee.mul(exchangeFeeRate ?? 0);
-							setQuoteCurrencyAmount(
-								truncateNumbers(baseCurrencyAmountNoFee.add(fee), DEFAULT_CRYPTO_DECIMALS)
-							);
-						}
-					}
-				}}
+				onBalanceClick={onBaseBalanceClick}
 				onCurrencySelect={
 					allowBaseCurrencySelection ? () => setOpenModal('base-select') : undefined
 				}
@@ -100,10 +71,8 @@ const BaseCurrencyCard: React.FC<BaseCurrencyCardProps> = ({ allowBaseCurrencySe
 							quote: pair.quote === currencyKey ? null : pair.quote,
 						}));
 
-						if (quoteCurrencyKey != null) {
-							if (quoteCurrencyKey !== currencyKey) {
-								routeToMarketPair(currencyKey, quoteCurrencyKey);
-							}
+						if (quoteCurrencyKey != null && quoteCurrencyKey !== currencyKey) {
+							routeToMarketPair(currencyKey, quoteCurrencyKey);
 						} else {
 							routeToBaseCurrency(currencyKey);
 						}
