+++ sections/exchange/TradeCard/Cards/QuoteCurrencyCard.tsx
@@ -1,9 +1,7 @@
-import { wei } from '@synthetixio/wei';
-import { DEFAULT_CRYPTO_DECIMALS } from 'constants/defaults';
-import { useExchangeContext } from 'contexts/ExchangeContext';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
+
 import SelectCurrencyModal from 'sections/shared/modals/SelectCurrencyModal';
 import {
 	baseCurrencyKeyState,
@@ -12,9 +10,9 @@ import {
 	quoteCurrencyAmountState,
 	currencyPairState,
 } from 'store/exchange';
-import { truncateNumbers } from 'utils/formatters/number';
 import CurrencyCard from '../CurrencyCard';
 import { CurrencyKey } from 'constants/currency';
+import { useExchangeContext } from 'contexts/ExchangeContext';
 
 type QuoteCurrencyCardProps = {
 	allowQuoteCurrencySelection?: boolean;
@@ -25,7 +23,7 @@ const QuoteCurrencyCard: React.FC<QuoteCurrencyCardProps> = ({ allowQuoteCurrenc
 	const baseCurrencyKey = useRecoilValue(baseCurrencyKeyState);
 	const quoteCurrencyKey = useRecoilValue(quoteCurrencyKeyState);
 	const setBaseCurrencyAmount = useSetRecoilState(baseCurrencyAmountState);
-	const [quoteCurrencyAmount, setQuoteCurrencyAmount] = useRecoilState(quoteCurrencyAmountState);
+	const quoteCurrencyAmount = useRecoilValue(quoteCurrencyAmountState);
 	const setCurrencyPair = useSetRecoilState(currencyPairState);
 
 	const {
@@ -35,9 +33,9 @@ const QuoteCurrencyCard: React.FC<QuoteCurrencyCardProps> = ({ allowQuoteCurrenc
 		setOpenModal,
 		routeToMarketPair,
 		allTokensMap,
-		exchangeFeeRate,
 		quotePriceRate,
-		rate,
+		onQuoteCurrencyAmountChange,
+		onQuoteBalanceClick,
 	} = useExchangeContext();
 
 	return (
@@ -47,46 +45,9 @@ const QuoteCurrencyCard: React.FC<QuoteCurrencyCardProps> = ({ allowQuoteCurrenc
 				currencyKey={quoteCurrencyKey}
 				currencyName={quoteCurrencyKey ? allTokensMap[quoteCurrencyKey]?.name : null}
 				amount={quoteCurrencyAmount}
-				onAmountChange={async (value) => {
-					if (value === '') {
-						setQuoteCurrencyAmount('');
-						setBaseCurrencyAmount('');
-					} else {
-						setQuoteCurrencyAmount(value);
-						if (txProvider === 'synthetix' && baseCurrencyKey != null) {
-							const baseCurrencyAmountNoFee = wei(value).mul(rate);
-							const fee = baseCurrencyAmountNoFee.mul(exchangeFeeRate ?? 0);
-							setBaseCurrencyAmount(
-								truncateNumbers(baseCurrencyAmountNoFee.sub(fee), DEFAULT_CRYPTO_DECIMALS)
-							);
-						}
-					}
-				}}
+				onAmountChange={onQuoteCurrencyAmountChange}
 				walletBalance={quoteCurrencyBalance}
-				onBalanceClick={async () => {
-					if (quoteCurrencyBalance != null) {
-						if ((quoteCurrencyKey as string) === 'ETH') {
-							const ETH_TX_BUFFER = 0.1;
-							const balanceWithBuffer = quoteCurrencyBalance.sub(wei(ETH_TX_BUFFER));
-							setQuoteCurrencyAmount(
-								balanceWithBuffer.lt(0)
-									? '0'
-									: truncateNumbers(balanceWithBuffer, DEFAULT_CRYPTO_DECIMALS)
-							);
-						} else {
-							setQuoteCurrencyAmount(
-								truncateNumbers(quoteCurrencyBalance, DEFAULT_CRYPTO_DECIMALS)
-							);
-						}
-						if (txProvider === 'synthetix') {
-							const baseCurrencyAmountNoFee = quoteCurrencyBalance.mul(rate);
-							const fee = baseCurrencyAmountNoFee.mul(exchangeFeeRate ?? 0);
-							setBaseCurrencyAmount(
-								truncateNumbers(baseCurrencyAmountNoFee.sub(fee), DEFAULT_CRYPTO_DECIMALS)
-							);
-						}
-					}
-				}}
+				onBalanceClick={onQuoteBalanceClick}
 				onCurrencySelect={
 					allowQuoteCurrencySelection ? () => setOpenModal('quote-select') : undefined
 				}
